make FilteringPaginationIterator top-level, move to new package
