+++ versioned/spi/src/main/java/org/projectnessie/versioned/MetricsVersionStore.java
@@ -32,6 +32,7 @@ import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import org.projectnessie.model.CommitMeta;
 import org.projectnessie.model.Content;
+import org.projectnessie.versioned.paging.PaginationIterator;
 
 /** A {@link VersionStore} wrapper that publishes metrics via Micrometer. */
 @SuppressWarnings("MustBeClosedChecker")
